Seperated posts from the main postings page so it can be reused.
